Migrate from Mumbai to Amoy (#139)
This is not backwards compatible as I changed the name of the network.

Amoy is too different from Mumbai to do a gracefull, backwards
compatible migration, as contract addresses have changed for anything a
dev touches on-chain. The rename of the network in our SDK makes that
much more explicit for the developer.